Rename shazbot to team181 for submitting
